Remove the undefined shifts a sanitizer build reports
Reading a 32-bit field out of a PIM message shifted a byte straight into
the sign bit of an int: GET_HOSTLONG() and GET_NETLONG() on a Hello's
GenID or DR priority, or an Assert's metric, whenever the leading byte is
0x80 or more.  Cast each byte to uint32_t, which every caller already
hands the macros as a uint8_t.

The others are not reachable from the wire.  do_randomize() rotated its
fallback seed by the seed itself, a shift of up to 2^32 where 32 is
already undefined, and a shift of 32 - 0 when it was zero; rotate by the
low five bits and leave it alone when they are zero.  domask() and the
walk over a routing socket reply's address bits shifted a 1 into an int's
sign bit, and so did the vif and neighbour bitmap macros, which no
sanitizer run reached but vif 31 and neighbours 31 and 63 would.

UndefinedBehaviorSanitizer reported the first four during a keepalive run
on FreeBSD.  An ASan+UBSan build now walks run all with no report at all,
nineteen of nineteen on both FreeBSD and Linux, with the option reaching
pimd and the log verified by a deliberate 1 << 31 beside it.

+++ src/pimd.h
@@ -508,9 +508,9 @@ typedef struct pim_jp_encod_grp_ {
 #define GET_HOSTLONG(val, cp)                   \
         do {                                    \
                 uint32_t Xv;                    \
-                Xv  = (*(cp)++) << 24;          \
-                Xv |= (*(cp)++) << 16;          \
-                Xv |= (*(cp)++) <<  8;          \
+                Xv  = (uint32_t)(*(cp)++) << 24; \
+                Xv |= (uint32_t)(*(cp)++) << 16; \
+                Xv |= (uint32_t)(*(cp)++) <<  8; \
                 Xv |= *(cp)++;                  \
                 (val) = Xv;                     \
         } while (0)
@@ -530,9 +530,9 @@ typedef struct pim_jp_encod_grp_ {
         do {                                    \
                 uint32_t Xv;                    \
                 Xv  = *(cp)++;                  \
-                Xv |= (*(cp)++) <<  8;          \
-                Xv |= (*(cp)++) << 16;          \
-                Xv |= (*(cp)++) << 24;          \
+                Xv |= (uint32_t)(*(cp)++) <<  8; \
+                Xv |= (uint32_t)(*(cp)++) << 16; \
+                Xv |= (uint32_t)(*(cp)++) << 24; \
                 (val) = Xv;                     \
         } while (0)
